added scaling with scale factors 1.25 and 0.8

+++ bitviewer.h
@@ -20,7 +20,7 @@ private slots:
 	void on_pushButton_minus_clicked();
 
 private:
-	size_t scale_;
+	double scale_;
 	Ui::BitViewerMainWindow ui;
 };
